64-bit compatibility
git-svn-id: https://welly.googlecode.com/svn/trunk@655 ace27a1a-5e51-0410-a197-e9aa1ca61478

+++ DBPrefsWindowController.m
@@ -89,7 +89,7 @@ static DBPrefsWindowController *_sharedPrefsWindowController = nil;
     /* Select the default client */
     CFStringRef defaultHandler = LSCopyDefaultHandlerForURLScheme((CFStringRef) scheme);
     if (defaultHandler) {
-        int index = [button indexOfItemWithRepresentedObject: (NSString *) defaultHandler];
+        NSInteger index = [button indexOfItemWithRepresentedObject: (NSString *) defaultHandler];
         if (index != NSNotFound) 
             [button selectItemAtIndex: index];
         CFRelease(defaultHandler);
